// apps/desktop/src/types/generated/ConflictIgnoreAction.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ConflictIgnoreAction =
  | { type: "ignorePair"; modA: string; modB: string }
  | { type: "unignorePair"; modA: string; modB: string }
  | { type: "ignorePairFile"; modA: string; modB: string; path: string }
  | { type: "ignoreMod"; modId: string }
  | { type: "unignoreMod"; modId: string }
  | { type: "clearAll" };